+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Removed `json` flag from `app download` command
+- Download progress bars now show app name instead of url leaf
 
 ## [1.15.0] - 2024-03-11